+++ cpp/core/CMakeLists.txt
@@ -184,8 +184,8 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-deprecated-declarations -Wno-attrib
 set(SPARK_COLUMNAR_PLUGIN_SRCS
         ${SUBSTRAIT_PROTO_SRCS}
         ${GLUTEN_PROTO_SRCS}
+        compute/ExecutionCtx.cc
         compute/ProtobufUtils.cc
-        compute/Backend.cc
         compute/ResultIterator.cc
         config/GlutenConfig.cc
         jni/JniWrapper.cc
