+++ cpp/core/tests/CMakeLists.txt
@@ -12,7 +12,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-add_test_case(exec_backend_test SOURCES BackendTest.cc)
+add_test_case(execution_ctx_test SOURCES ExecutionCtxTest.cc)
 
 if(ENABLE_HBM)
   add_test_case(hbw_allocator_test SOURCES HbwAllocatorTest.cc)
