+++ cpp/velox/benchmarks/GenericBenchmark.cc
@@ -27,7 +27,7 @@
 #include "BatchStreamIterator.h"
 #include "BatchVectorIterator.h"
 #include "BenchmarkUtils.h"
-#include "compute/VeloxBackend.h"
+#include "compute/VeloxExecutionCtx.h"
 #include "compute/VeloxPlanConverter.h"
 #include "config/GlutenConfig.h"
 #include "shuffle/LocalPartitionWriter.h"
@@ -133,7 +133,7 @@ auto BM_Generic = [](::benchmark::State& state,
   WriterMetrics writerMetrics{};
 
   for (auto _ : state) {
-    auto backend = gluten::createBackend();
+    auto executionCtx = gluten::createExecutionCtx();
     std::vector<std::shared_ptr<gluten::ResultIterator>> inputIters;
     std::vector<BatchIterator*> inputItersRaw;
     if (!inputFiles.empty()) {
@@ -147,9 +147,10 @@ auto BM_Generic = [](::benchmark::State& state,
           });
     }
 
-    backend->parsePlan(reinterpret_cast<uint8_t*>(plan.data()), plan.size());
-    auto resultIter = backend->getResultIterator(memoryManager.get(), "/tmp/test-spill", std::move(inputIters), conf);
-    auto veloxPlan = std::dynamic_pointer_cast<gluten::VeloxBackend>(backend)->getVeloxPlan();
+    executionCtx->parsePlan(reinterpret_cast<uint8_t*>(plan.data()), plan.size());
+    auto resultIter =
+        executionCtx->getResultIterator(memoryManager.get(), "/tmp/test-spill", std::move(inputIters), conf);
+    auto veloxPlan = std::dynamic_pointer_cast<gluten::VeloxExecutionCtx>(executionCtx)->getVeloxPlan();
     if (FLAGS_with_shuffle) {
       int64_t shuffleWriteTime;
       TIME_NANO_START(shuffleWriteTime);
