+++ cpp/velox/benchmarks/QueryBenchmark.cc
@@ -16,7 +16,7 @@
  */
 
 #include <benchmark/benchmark.h>
-#include <compute/VeloxBackend.h>
+#include <compute/VeloxExecutionCtx.h>
 
 #include "BenchmarkUtils.h"
 #include "compute/VeloxPlanConverter.h"
@@ -35,7 +35,7 @@ const std::string getFilePath(const std::string& fileName) {
 // Used by unit test and benchmark.
 std::shared_ptr<ResultIterator> getResultIterator(
     std::shared_ptr<velox::memory::MemoryPool> veloxPool,
-    std::shared_ptr<Backend> backend,
+    std::shared_ptr<ExecutionCtx> executionCtx,
     const std::vector<std::shared_ptr<SplitInfo>>& setScanInfos,
     std::shared_ptr<const facebook::velox::core::PlanNode>& veloxPlan) {
   auto ctxPool = veloxPool->addAggregateChild(
@@ -45,7 +45,7 @@ std::shared_ptr<ResultIterator> getResultIterator(
   std::unordered_map<std::string, std::string> sessionConf = {};
   auto veloxPlanConverter =
       std::make_unique<VeloxPlanConverter>(inputIter, defaultLeafVeloxMemoryPool().get(), sessionConf);
-  veloxPlan = veloxPlanConverter->toVeloxPlan(backend->getPlan());
+  veloxPlan = veloxPlanConverter->toVeloxPlan(executionCtx->getPlan());
 
   // In test, use setScanInfos to replace the one got from Substrait.
   std::vector<std::shared_ptr<SplitInfo>> scanInfos;
@@ -53,7 +53,7 @@ std::shared_ptr<ResultIterator> getResultIterator(
   std::vector<velox::core::PlanNodeId> streamIds;
 
   // Separate the scan ids and stream ids, and get the scan infos.
-  VeloxBackend::getInfoAndIds(
+  VeloxExecutionCtx::getInfoAndIds(
       veloxPlanConverter->splitInfos(), veloxPlan->leafPlanNodeIds(), scanInfos, scanIds, streamIds);
 
   auto wholestageIter = std::make_unique<WholeStageResultIteratorFirstStage>(
@@ -63,9 +63,9 @@ std::shared_ptr<ResultIterator> getResultIterator(
       setScanInfos,
       streamIds,
       "/tmp/test-spill",
-      backend->getConfMap(),
-      backend->getSparkTaskInfo());
-  return std::make_shared<ResultIterator>(std::move(wholestageIter), backend);
+      executionCtx->getConfMap(),
+      executionCtx->getSparkTaskInfo());
+  return std::make_shared<ResultIterator>(std::move(wholestageIter), executionCtx);
 }
 
 auto BM = [](::benchmark::State& state,
@@ -90,12 +90,12 @@ auto BM = [](::benchmark::State& state,
 
   for (auto _ : state) {
     state.PauseTiming();
-    auto backend = std::dynamic_pointer_cast<gluten::VeloxBackend>(gluten::createBackend());
+    auto executionCtx = std::dynamic_pointer_cast<gluten::VeloxExecutionCtx>(gluten::createExecutionCtx());
     state.ResumeTiming();
 
-    backend->parsePlan(reinterpret_cast<uint8_t*>(plan.data()), plan.size());
+    executionCtx->parsePlan(reinterpret_cast<uint8_t*>(plan.data()), plan.size());
     std::shared_ptr<const facebook::velox::core::PlanNode> veloxPlan;
-    auto resultIter = getResultIterator(veloxPool, backend, scanInfos, veloxPlan);
+    auto resultIter = getResultIterator(veloxPool, executionCtx, scanInfos, veloxPlan);
     auto outputSchema = toArrowSchema(veloxPlan->outputType(), defaultLeafVeloxMemoryPool().get());
     while (resultIter->hasNext()) {
       auto array = resultIter->next()->exportArrowArray();
