+++ cpp/velox/compute/VeloxBackend.cc
@@ -14,127 +14,320 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-#include "VeloxBackend.h"
 #include <filesystem>
 
-#include "arrow/c/bridge.h"
-#include "compute/Backend.h"
-#include "compute/ResultIterator.h"
-#include "compute/VeloxPlanConverter.h"
-#include "config/GlutenConfig.h"
-#include "operators/serializer/VeloxRowToColumnarConverter.h"
-#include "shuffle/VeloxShuffleWriter.h"
+#include "VeloxBackend.h"
+
+#include <folly/executors/IOThreadPoolExecutor.h>
+
+#include "operators/functions/RegistrationAllFunctions.h"
+#include "operators/plannodes/RowVectorStream.h"
+#ifdef GLUTEN_ENABLE_QAT
+#include "utils/qat/QatCodec.h"
+#endif
+#ifdef GLUTEN_ENABLE_IAA
+#include "utils/qpl/qpl_codec.h"
+#endif
+#include "utils/exception.h"
 #include "velox/common/file/FileSystems.h"
+#include "velox/serializers/PrestoSerializer.h"
+#ifdef ENABLE_HDFS
+#include "velox/connectors/hive/storage_adapters/hdfs/HdfsFileSystem.h"
+#endif
+#ifdef ENABLE_S3
+#include "velox/connectors/hive/storage_adapters/s3fs/S3FileSystem.h"
+#endif
+#include "jni/JniFileSystem.h"
+#include "udf/UdfLoader.h"
+#include "utils/ConfigExtractor.h"
+#include "velox/common/memory/MmapAllocator.h"
+#include "velox/connectors/hive/HiveConnector.h"
+#include "velox/dwio/dwrf/reader/DwrfReader.h"
+#include "velox/dwio/parquet/RegisterParquetReader.h"
+
+DECLARE_int32(split_preload_per_driver);
+DECLARE_bool(SkipRowSortInWindowOp);
+DECLARE_bool(velox_exception_user_stacktrace_enabled);
 
 using namespace facebook;
 
-namespace gluten {
-
 namespace {
 
-void printSessionConf(const std::unordered_map<std::string, std::string>& conf) {
+const std::string kEnableUserExceptionStacktrace =
+    "spark.gluten.sql.columnar.backend.velox.enableUserExceptionStacktrace";
+const std::string kEnableUserExceptionStacktraceDefault = "true";
+
+const std::string kHiveConnectorId = "test-hive";
+const std::string kVeloxCacheEnabled = "spark.gluten.sql.columnar.backend.velox.cacheEnabled";
+
+// memory cache
+const std::string kVeloxMemCacheSize = "spark.gluten.sql.columnar.backend.velox.memCacheSize";
+const std::string kVeloxMemCacheSizeDefault = "1073741824";
+
+// ssd cache
+const std::string kVeloxSsdCacheSize = "spark.gluten.sql.columnar.backend.velox.ssdCacheSize";
+const std::string kVeloxSsdCacheSizeDefault = "1073741824";
+const std::string kVeloxSsdCachePath = "spark.gluten.sql.columnar.backend.velox.ssdCachePath";
+const std::string kVeloxSsdCachePathDefault = "/tmp/";
+const std::string kVeloxSsdCacheShards = "spark.gluten.sql.columnar.backend.velox.ssdCacheShards";
+const std::string kVeloxSsdCacheShardsDefault = "1";
+const std::string kVeloxSsdCacheIOThreads = "spark.gluten.sql.columnar.backend.velox.ssdCacheIOThreads";
+const std::string kVeloxSsdCacheIOThreadsDefault = "1";
+const std::string kVeloxSsdODirectEnabled = "spark.gluten.sql.columnar.backend.velox.ssdODirect";
+
+const std::string kVeloxIOThreads = "spark.gluten.sql.columnar.backend.velox.IOThreads";
+const std::string kVeloxIOThreadsDefault = "0";
+
+const std::string kVeloxSplitPreloadPerDriver = "spark.gluten.sql.columnar.backend.velox.SplitPreloadPerDriver";
+const std::string kVeloxSplitPreloadPerDriverDefault = "2";
+
+// udf
+const std::string kVeloxUdfLibraryPaths = "spark.gluten.sql.columnar.backend.velox.udfLibraryPaths";
+
+// spill
+const std::string kMaxSpillFileSize = "spark.gluten.sql.columnar.backend.velox.maxSpillFileSize";
+const std::string kMaxSpillFileSizeDefault = std::to_string(20L * 1024 * 1024);
+
+} // namespace
+
+namespace gluten {
+
+void VeloxBackend::printConf(const std::unordered_map<std::string, std::string>& conf) {
   std::ostringstream oss;
-  oss << "session conf = {\n";
+  oss << "STARTUP: VeloxBackend conf = {\n";
   for (auto& [k, v] : conf) {
-    oss << " {" << k << " = " << v << "}\n";
+    oss << " {" << k << ", " << v << "}\n";
   }
   oss << "}\n";
   LOG(INFO) << oss.str();
 }
 
-} // namespace
-
-VeloxBackend::VeloxBackend(const std::unordered_map<std::string, std::string>& confMap) : Backend(confMap) {}
-
-void VeloxBackend::getInfoAndIds(
-    const std::unordered_map<velox::core::PlanNodeId, std::shared_ptr<SplitInfo>>& splitInfoMap,
-    const std::unordered_set<velox::core::PlanNodeId>& leafPlanNodeIds,
-    std::vector<std::shared_ptr<SplitInfo>>& scanInfos,
-    std::vector<velox::core::PlanNodeId>& scanIds,
-    std::vector<velox::core::PlanNodeId>& streamIds) {
-  for (const auto& leafPlanNodeId : leafPlanNodeIds) {
-    auto it = splitInfoMap.find(leafPlanNodeId);
-    if (it == splitInfoMap.end()) {
-      throw std::runtime_error("Could not find leafPlanNodeId.");
+void VeloxBackend::init(const std::unordered_map<std::string, std::string>& conf) {
+  // In spark, planner takes care the parititioning and sorting, so the rows are sorted.
+  // There is no need to sort the rows in window op again.
+  FLAGS_SkipRowSortInWindowOp = true;
+  // Set velox_exception_user_stacktrace_enabled.
+  {
+    auto got = conf.find(kEnableUserExceptionStacktrace);
+    std::string enableUserExceptionStacktrace = kEnableUserExceptionStacktraceDefault;
+    if (got != conf.end()) {
+      enableUserExceptionStacktrace = got->second;
     }
-    auto splitInfo = it->second;
-    if (splitInfo->isStream) {
-      streamIds.emplace_back(leafPlanNodeId);
-    } else {
-      scanInfos.emplace_back(splitInfo);
-      scanIds.emplace_back(leafPlanNodeId);
-    }
+    FLAGS_velox_exception_user_stacktrace_enabled = (enableUserExceptionStacktrace == "true");
   }
-}
 
-std::shared_ptr<ResultIterator> VeloxBackend::getResultIterator(
-    MemoryManager* memoryManager,
-    const std::string& spillDir,
-    const std::vector<std::shared_ptr<ResultIterator>>& inputs,
-    const std::unordered_map<std::string, std::string>& sessionConf) {
-#ifdef GLUTEN_PRINT_DEBUG
-  printSessionConf(sessionConf);
+  // Setup and register.
+  velox::filesystems::registerLocalFileSystem();
+  initJolFilesystem(conf);
+
+#ifdef ENABLE_HDFS
+  velox::filesystems::registerHdfsFileSystem();
 #endif
-  if (inputs.size() > 0) {
-    inputIters_ = std::move(inputs);
+
+  std::unordered_map<std::string, std::string> configurationValues;
+#ifdef ENABLE_S3
+  velox::filesystems::registerS3FileSystem();
+
+  std::string awsAccessKey = conf.at("spark.hadoop.fs.s3a.access.key");
+  std::string awsSecretKey = conf.at("spark.hadoop.fs.s3a.secret.key");
+  std::string awsEndpoint = conf.at("spark.hadoop.fs.s3a.endpoint");
+  std::string sslEnabled = conf.at("spark.hadoop.fs.s3a.connection.ssl.enabled");
+  std::string pathStyleAccess = conf.at("spark.hadoop.fs.s3a.path.style.access");
+  std::string useInstanceCredentials = conf.at("spark.hadoop.fs.s3a.use.instance.credentials");
+  std::string iamRole = conf.at("spark.hadoop.fs.s3a.iam.role");
+  std::string iamRoleSessionName = conf.at("spark.hadoop.fs.s3a.iam.role.session.name");
+
+  const char* envAwsAccessKey = std::getenv("AWS_ACCESS_KEY_ID");
+  if (envAwsAccessKey != nullptr) {
+    awsAccessKey = std::string(envAwsAccessKey);
+  }
+  const char* envAwsSecretKey = std::getenv("AWS_SECRET_ACCESS_KEY");
+  if (envAwsSecretKey != nullptr) {
+    awsSecretKey = std::string(envAwsSecretKey);
+  }
+  const char* envAwsEndpoint = std::getenv("AWS_ENDPOINT");
+  if (envAwsEndpoint != nullptr) {
+    awsEndpoint = std::string(envAwsEndpoint);
   }
 
-  auto veloxPool = getAggregateVeloxPool(memoryManager);
-
-  VeloxPlanConverter veloxPlanConverter(inputIters_, getLeafVeloxPool(memoryManager).get(), sessionConf);
-  veloxPlan_ = veloxPlanConverter.toVeloxPlan(substraitPlan_);
-
-  // Scan node can be required.
-  std::vector<std::shared_ptr<SplitInfo>> scanInfos;
-  std::vector<velox::core::PlanNodeId> scanIds;
-  std::vector<velox::core::PlanNodeId> streamIds;
-
-  // Separate the scan ids and stream ids, and get the scan infos.
-  getInfoAndIds(veloxPlanConverter.splitInfos(), veloxPlan_->leafPlanNodeIds(), scanInfos, scanIds, streamIds);
-
-  if (scanInfos.size() == 0) {
-    // Source node is not required.
-    auto wholestageIter = std::make_unique<WholeStageResultIteratorMiddleStage>(
-        veloxPool, veloxPlan_, streamIds, spillDir, sessionConf, taskInfo_);
-    return std::make_shared<ResultIterator>(std::move(wholestageIter), shared_from_this());
+  std::unordered_map<std::string, std::string> s3Config({});
+  if (useInstanceCredentials == "true") {
+    s3Config.insert({
+        {"hive.s3.use-instance-credentials", useInstanceCredentials},
+    });
+  } else if (!iamRole.empty()) {
+    s3Config.insert({
+        {"hive.s3.iam-role", iamRole},
+    });
+    if (!iamRoleSessionName.empty()) {
+      s3Config.insert({
+          {"hive.s3.iam-role-session-name", iamRoleSessionName},
+      });
+    }
   } else {
-    auto wholestageIter = std::make_unique<WholeStageResultIteratorFirstStage>(
-        veloxPool, veloxPlan_, scanIds, scanInfos, streamIds, spillDir, sessionConf, taskInfo_);
-    return std::make_shared<ResultIterator>(std::move(wholestageIter), shared_from_this());
+    s3Config.insert({
+        {"hive.s3.aws-access-key", awsAccessKey},
+        {"hive.s3.aws-secret-key", awsSecretKey},
+    });
+  }
+  // Only need to set s3 endpoint when not use instance credentials.
+  if (useInstanceCredentials != "true") {
+    s3Config.insert({
+        {"hive.s3.endpoint", awsEndpoint},
+    });
+  }
+  s3Config.insert({
+      {"hive.s3.ssl.enabled", sslEnabled},
+      {"hive.s3.path-style-access", pathStyleAccess},
+  });
+
+  configurationValues.merge(s3Config);
+#endif
+
+  initCache(conf);
+  initIOExecutor(conf);
+
+#ifdef GLUTEN_PRINT_DEBUG
+  printConf(conf);
+#endif
+
+  auto properties = std::make_shared<const velox::core::MemConfig>(configurationValues);
+  velox::connector::registerConnectorFactory(std::make_shared<velox::connector::hive::HiveConnectorFactory>());
+  auto hiveConnector =
+      velox::connector::getConnectorFactory(velox::connector::hive::HiveConnectorFactory::kHiveConnectorName)
+          ->newConnector(kHiveConnectorId, properties, ioExecutor_.get());
+
+  registerConnector(hiveConnector);
+  velox::parquet::registerParquetReaderFactory(velox::parquet::ParquetReaderType::NATIVE);
+  velox::dwrf::registerDwrfReaderFactory();
+  velox::dwrf::registerOrcReaderFactory();
+  // Register Velox functions
+  registerAllFunctions();
+  if (!facebook::velox::isRegisteredVectorSerde()) {
+    // serde, for spill
+    facebook::velox::serializer::presto::PrestoVectorSerde::registerVectorSerde();
+  }
+  velox::exec::Operator::registerOperator(std::make_unique<RowVectorStreamOperatorTranslator>());
+
+  initUdf(conf);
+}
+
+velox::memory::MemoryAllocator* VeloxBackend::getAsyncDataCache() const {
+  return asyncDataCache_.get();
+}
+
+// JNI-or-local filesystem, for spilling-to-heap if we have extra JVM heap spaces
+void VeloxBackend::initJolFilesystem(const std::unordered_map<std::string, std::string>& conf) {
+  int64_t maxSpillFileSize = std::stol(kMaxSpillFileSizeDefault);
+  auto got = conf.find(kMaxSpillFileSize);
+  if (got != conf.end()) {
+    maxSpillFileSize = std::stol(got->second);
+  }
+  // FIMXE It's known that if spill compression is disabled, the actual spill file size may
+  //   in crease beyond this limit a little (maximum 64 rows which is by default
+  //   one compression page)
+  gluten::registerJolFileSystem(maxSpillFileSize);
+}
+
+void VeloxBackend::initCache(const std::unordered_map<std::string, std::string>& conf) {
+  auto key = conf.find(kVeloxCacheEnabled);
+  if (key != conf.end() && boost::algorithm::to_lower_copy(conf.at(kVeloxCacheEnabled)) == "true") {
+    FLAGS_ssd_odirect = true;
+    if (conf.find(kVeloxSsdODirectEnabled) != conf.end() &&
+        boost::algorithm::to_lower_copy(conf.at(kVeloxSsdODirectEnabled)) == "false") {
+      FLAGS_ssd_odirect = false;
+    }
+    uint64_t memCacheSize = std::stol(kVeloxMemCacheSizeDefault);
+    uint64_t ssdCacheSize = std::stol(kVeloxSsdCacheSizeDefault);
+    int32_t ssdCacheShards = std::stoi(kVeloxSsdCacheShardsDefault);
+    int32_t ssdCacheIOThreads = std::stoi(kVeloxSsdCacheIOThreadsDefault);
+    std::string ssdCachePathPrefix = kVeloxSsdCachePathDefault;
+    for (auto& [k, v] : conf) {
+      if (k == kVeloxMemCacheSize)
+        memCacheSize = std::stol(v);
+      if (k == kVeloxSsdCacheSize)
+        ssdCacheSize = std::stol(v);
+      if (k == kVeloxSsdCacheShards)
+        ssdCacheShards = std::stoi(v);
+      if (k == kVeloxSsdCachePath)
+        ssdCachePathPrefix = v;
+      if (k == kVeloxSsdCacheIOThreads)
+        ssdCacheIOThreads = std::stoi(v);
+    }
+    cachePathPrefix_ = ssdCachePathPrefix;
+    cacheFilePrefix_ = getCacheFilePrefix();
+    std::string ssdCachePath = ssdCachePathPrefix + "/" + cacheFilePrefix_;
+    ssdCacheExecutor_ = std::make_unique<folly::IOThreadPoolExecutor>(ssdCacheIOThreads);
+    auto ssd =
+        std::make_unique<velox::cache::SsdCache>(ssdCachePath, ssdCacheSize, ssdCacheShards, ssdCacheExecutor_.get());
+
+    std::error_code ec;
+    const std::filesystem::space_info si = std::filesystem::space(ssdCachePathPrefix, ec);
+    if (si.available < ssdCacheSize) {
+      VELOX_FAIL(
+          "not enough space for ssd cache in " + ssdCachePath + " cache size: " + std::to_string(ssdCacheSize) +
+          "free space: " + std::to_string(si.available))
+    }
+
+    velox::memory::MmapAllocator::Options options;
+    options.capacity = memCacheSize;
+    auto allocator = std::make_shared<velox::memory::MmapAllocator>(options);
+    if (ssdCacheSize == 0) {
+      LOG(INFO) << "AsyncDataCache will do memory caching only as ssd cache size is 0";
+      asyncDataCache_ = std::make_shared<velox::cache::AsyncDataCache>(allocator, memCacheSize, nullptr);
+    } else {
+      asyncDataCache_ = std::make_shared<velox::cache::AsyncDataCache>(allocator, memCacheSize, std::move(ssd));
+    }
+
+    VELOX_CHECK_NOT_NULL(dynamic_cast<velox::cache::AsyncDataCache*>(asyncDataCache_.get()))
+    LOG(INFO) << "STARTUP: Using AsyncDataCache memory cache size: " << memCacheSize
+              << ", ssdCache prefix: " << ssdCachePath << ", ssdCache size: " << ssdCacheSize
+              << ", ssdCache shards: " << ssdCacheShards << ", ssdCache IO threads: " << ssdCacheIOThreads;
   }
 }
 
-std::shared_ptr<ColumnarToRowConverter> VeloxBackend::getColumnar2RowConverter(MemoryManager* memoryManager) {
-  auto ctxVeloxPool = getLeafVeloxPool(memoryManager);
-  return std::make_shared<VeloxColumnarToRowConverter>(ctxVeloxPool);
+void VeloxBackend::initIOExecutor(const std::unordered_map<std::string, std::string>& conf) {
+  int32_t ioThreads = std::stoi(getConfigValue(conf, kVeloxIOThreads, kVeloxIOThreadsDefault));
+  int32_t splitPreloadPerDriver =
+      std::stoi(getConfigValue(conf, kVeloxSplitPreloadPerDriver, kVeloxSplitPreloadPerDriverDefault));
+  if (ioThreads > 0) {
+    ioExecutor_ = std::make_unique<folly::IOThreadPoolExecutor>(ioThreads);
+    FLAGS_split_preload_per_driver = splitPreloadPerDriver;
+  }
+
+  if (splitPreloadPerDriver > 0 && ioThreads > 0) {
+    LOG(INFO) << "STARTUP: Using split preloading, Split preload per driver: " << splitPreloadPerDriver
+              << ", IO threads: " << ioThreads;
+  }
 }
 
-std::shared_ptr<RowToColumnarConverter> VeloxBackend::getRowToColumnarConverter(
-    MemoryManager* memoryManager,
-    struct ArrowSchema* cSchema) {
-  auto ctxVeloxPool = getLeafVeloxPool(memoryManager);
-  return std::make_shared<VeloxRowToColumnarConverter>(cSchema, ctxVeloxPool);
+void VeloxBackend::initUdf(const std::unordered_map<std::string, std::string>& conf) {
+  auto got = conf.find(kVeloxUdfLibraryPaths);
+  if (got != conf.end() && !got->second.empty()) {
+    auto udfLoader = gluten::UdfLoader::getInstance();
+    udfLoader->loadUdfLibraries(got->second);
+    udfLoader->registerUdf();
+  }
 }
 
-std::shared_ptr<ShuffleWriter> VeloxBackend::createShuffleWriter(
-    int numPartitions,
-    std::shared_ptr<ShuffleWriter::PartitionWriterCreator> partitionWriterCreator,
-    const ShuffleWriterOptions& options,
-    MemoryManager* memoryManager) {
-  auto ctxPool = getLeafVeloxPool(memoryManager);
-  GLUTEN_ASSIGN_OR_THROW(
-      auto shuffle_writer,
-      VeloxShuffleWriter::create(numPartitions, std::move(partitionWriterCreator), std::move(options), ctxPool));
-  return shuffle_writer;
+void VeloxBackend::create(const std::unordered_map<std::string, std::string>& conf) {
+  std::lock_guard<std::mutex> lockGuard(mutex_);
+  if (instance_ != nullptr) {
+    assert(false);
+    throw gluten::GlutenException("VeloxBackend already set");
+  }
+  instance_.reset(new gluten::VeloxBackend(conf));
 }
 
-std::shared_ptr<ColumnarBatchSerializer> VeloxBackend::getColumnarBatchSerializer(
-    MemoryManager* memoryManager,
-    std::shared_ptr<arrow::MemoryPool> arrowPool,
-    struct ArrowSchema* cSchema) {
-  auto ctxVeloxPool = getLeafVeloxPool(memoryManager);
-  return std::make_shared<VeloxColumnarBatchSerializer>(arrowPool, ctxVeloxPool, cSchema);
+std::shared_ptr<VeloxBackend> VeloxBackend::get() {
+  std::lock_guard<std::mutex> lockGuard(mutex_);
+  if (instance_ == nullptr) {
+    LOG(INFO) << "VeloxBackend not set, using default VeloxBackend instance. This should only happen in test code.";
+    static const std::unordered_map<std::string, std::string> kEmptyConf;
+    static std::shared_ptr<VeloxBackend> defaultInstance{new gluten::VeloxBackend(kEmptyConf)};
+    return defaultInstance;
+  }
+  return instance_;
 }
 
 } // namespace gluten
